move queueserver_agent testcases to new test module
